feat: add types for TODOS
