+++ news-api/src/entities/todos/model/types.ts
@@ -1,0 +1,13 @@
+export interface Todos {
+    userId: number;
+    id: number;
+    title: string;
+    completed: boolean;
+}
+
+export interface FilterTodos {
+    limit?: number;
+    userId?: number;
+    completed?: boolean;
+    page?: number;
+}
